requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
psycopg2-binary==2.9.9
pgvector==0.3.5
alembic==1.13.0
pydantic-settings==2.5.0
langchain==0.3.25
langchain-openai==0.3.12
langgraph==0.4.1
serpapi==0.1.5
pymupdf==1.24.0
httpx==0.27.0
python-dotenv==1.0.1
pytest==8.3.0
pytest-asyncio==0.24.0
